Add behavior that moves a window around when dragging a tab item to float it into a new window. The position of the window updates based on the relative position of the mouse when the TabItem started to be dragged.
Need to address an issue where the Mouse is not captured outside of the window. If you move the mouse quick enough you can get outside the bounds of the window, causing MouseMove events to be ignored until the mouse is over the window again. Thought that CaptureMouse() addressed this, but turns out I was just lucky.

GetParent, GetAncestor, and GetChildrenOfType now specify "Visual" as the element needs to be loaded for the VisualTreeHelper to work.